feat: Add order status and query endpoints
- Add GET /api/v1/orders/{order_id} endpoint for single order status
- Add GET /api/v1/orders endpoint for listing orders with filters
- Add OrderStatus enum (Pending, Active, Partial, Filled, Canceled)
- Add OrderTimeInForce enum (GTC, IOC, FOK, GTD)
- Add OrderInfo, OrderStatusResponse, OrderListResponse models
- Add OrderListQuery for filtering by underlying, status, side
- Support pagination with limit and offset parameters
- Add DashMap-based order storage in AppState
- Update price types from u64 to u128 for orderbook-rs 0.5 compatibility
- Add 6 new unit tests for order status endpoints
- Update dependencies to use local option-chain-orderbook 0.3.0
- Update orderbook-rs to 0.5 with special_orders feature

+++ src/market_maker/quoter.rs
@@ -7,9 +7,9 @@ use optionstratlib::{ExpirationDate, OptionStyle};
 #[derive(Debug, Clone)]
 pub struct QuoteParams {
     /// Bid price in cents.
-    pub bid_price: u64,
+    pub bid_price: u128,
     /// Ask price in cents.
-    pub ask_price: u64,
+    pub ask_price: u128,
     /// Bid size.
     pub bid_size: u64,
     /// Ask size.
@@ -96,9 +96,9 @@ impl Quoter {
         };
 
         let bid_price =
-            (theo_cents as i64 - half_spread_cents as i64 + bid_adjustment).max(1) as u64;
+            (theo_cents as i64 - half_spread_cents as i64 + bid_adjustment).max(1) as u128;
         let ask_price = (theo_cents as i64 + half_spread_cents as i64 + ask_adjustment)
-            .max(bid_price as i64 + 1) as u64;
+            .max(bid_price as i64 + 1) as u128;
 
         // Calculate sizes with scalar
         let base_size = (self.base_size as f64 * input.size_scalar).max(1.0) as u64;
